Pull correct apiKey from page source.
The Weather Underground home page contains many API keys to pull in
information to fill its forecasts, current conditions, etc. Prior to
this commit, the library grabbed the first key it found.

This commit searches for the key used to find

    "v2/observations/current"

which is the key we want to use. Also, rather than starting at offset
7 in the found string, this commit uses the `.captures()` method to
return the string in the grouping delimiters.